dont display rainbow grades if displaysummary is not active
